NuGet v1.1.0, fix for cases where chunk read/write methods are modifying chunks (encrypt, compress, etc), methods for object metadata

+++ DedupeCore/Common.cs
@@ -191,10 +191,10 @@ namespace WatsonDedupe
             return ret;
         }
 
-        public static byte[] InitBytes(int count, byte val)
+        public static byte[] InitBytes(long count, byte val)
         {
             byte[] ret = new byte[count];
-            for (int i = 0; i < ret.Length; i++)
+            for (long i = 0; i < ret.Length; i++)
             {
                 ret[i] = val;
             }
